+++ kibot/gs.py
@@ -69,6 +69,8 @@ class GS(object):
     global_from_cli = {}
     global_output = None
     global_variant = None
+    global_kiauto_wait_start = None
+    global_kiauto_time_out_scale = None
     global_opts_class = None
     test_boolean = True