+++ kibot/pre_drc.py
@@ -8,7 +8,7 @@ from .macros import macros, pre_class  # noqa: F401
 from .error import (KiPlotConfigurationError)
 from .gs import (GS)
 from .optionable import Optionable
-from .kiplot import check_script, exec_with_retry, load_board
+from .kiplot import check_script, exec_with_retry, load_board, add_time_out_options
 from .misc import (CMD_PCBNEW_RUN_DRC, URL_PCBNEW_RUN_DRC, DRC_ERROR)
 from .log import (get_logger)
 
@@ -45,6 +45,7 @@ class Run_DRC(BasePreFlight):  # noqa: F821
         if GS.debug_enabled:
             cmd.insert(1, '-vv')
             cmd.insert(1, '-r')
+        cmd = add_time_out_options(cmd)
         logger.info('- Running the DRC')
         ret = exec_with_retry(cmd)
         if ret: