+++ kibot/pre_erc.py
@@ -7,7 +7,7 @@ from sys import (exit)
 from .macros import macros, pre_class  # noqa: F401
 from .gs import (GS)
 from .optionable import Optionable
-from .kiplot import check_eeschema_do, exec_with_retry, load_sch
+from .kiplot import check_eeschema_do, exec_with_retry, load_sch, add_time_out_options
 from .error import (KiPlotConfigurationError)
 from .misc import (CMD_EESCHEMA_DO, ERC_ERROR)
 from .log import (get_logger)
@@ -45,6 +45,7 @@ class Run_ERC(BasePreFlight):  # noqa: F821
         if GS.debug_enabled:
             cmd.insert(1, '-vv')
             cmd.insert(1, '-r')
+        cmd = add_time_out_options(cmd)
         logger.info('- Running the ERC')
         ret = exec_with_retry(cmd)
         if ret: