+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [PCBWay](https://www.pcbway.com)
 - Support for ZIP/TAR/RAR generation.
 - Makefile generation.
+- KiAuto time-out control.
 
 ### Changed
 - Now the default output name applies to the DRC and ERC report names.